requirements.txt
alembic==1.13.2
annotated-types==0.7.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.0
cycler==0.12.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
greenlet==3.1.1
h2o==3.46.0.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.2
pandas==2.2.2
pillow==11.0.0
polars==1.21.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyarrow==19.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
SQLAlchemy==2.0.32
tabulate==0.9.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.4
